Electrical Lines Refactor
