+++ Slider/Assets/Prefabs/Map/Functional/Factory/Pilon.prefab
@@ -9,6 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 6837428105743242782}
+  - component: {fileID: 1726755163970623396}
   - component: {fileID: 6837428105743242784}
   m_Layer: 0
   m_Name: Pilon
@@ -27,12 +28,39 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 2778206746076517738}
   - {fileID: 8091606999308705072}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1726755163970623396
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6837428105743242786}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 82d6338f302a35d4d8a501dee3ec9f09, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  nodeType: 0
+  neighbors: []
+  invertSignal: 0
+  debugAsPoweredOn: 0
+  OnPowered:
+    m_PersistentCalls:
+      m_Calls: []
+  OnPoweredOn:
+    m_PersistentCalls:
+      m_Calls: []
+  OnPoweredOff:
+    m_PersistentCalls:
+      m_Calls: []
+  isConductiveItem: 0
 --- !u!114 &6837428105743242784
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -45,19 +73,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 82053ce22e7fd8d47a3645089ed5ee8d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  nodeType: 0
-  powerRefs: 0
-  neighbors: []
-  invertSignal: 0
-  OnPowered:
-    m_PersistentCalls:
-      m_Calls: []
-  OnPoweredOn:
-    m_PersistentCalls:
-      m_Calls: []
-  OnPoweredOff:
-    m_PersistentCalls:
-      m_Calls: []
+  eNode: {fileID: 1726755163970623396}
   lineStart: {fileID: 8091606999308705072}
   anim: {fileID: 8809015012945962431}
 --- !u!1 &8736454447659889831
@@ -90,6 +106,7 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0.1248, y: -0.1875, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 6837428105743242782}
   m_RootOrder: 0
@@ -105,6 +122,7 @@ SpriteRenderer:
   m_CastShadows: 0
   m_ReceiveShadows: 0
   m_DynamicOccludee: 1
+  m_StaticShadowCaster: 0
   m_MotionVectors: 1
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
@@ -132,7 +150,7 @@ SpriteRenderer:
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 529684183
-  m_SortingLayer: 2
+  m_SortingLayer: 6
   m_SortingOrder: 0
   m_Sprite: {fileID: -3124052884837581905, guid: fb5b2eb6132030c42958bb86a82ab4ed, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
@@ -147,7 +165,7 @@ SpriteRenderer:
   m_SpriteSortPoint: 0
 --- !u!95 &8809015012945962431
 Animator:
-  serializedVersion: 3
+  serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
@@ -160,6 +178,7 @@ Animator:
   m_UpdateMode: 0
   m_ApplyRootMotion: 0
   m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
   m_WarningMessage: 
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
@@ -242,6 +261,7 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0.118, y: -0.135, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 6837428105743242782}
   m_RootOrder: 1
