+++ Slider/Assets/Scripts/Map/Factory/BaseComponents/ConductiveElectricalNode.cs
@@ -7,6 +7,21 @@ public class ConductiveElectricalNode : ElectricalNode
     [Header("Conductive Electrical Node")]
     [SerializeField] private bool isConductiveItem;
 
+    public struct NodeEventArgs
+    {
+        public ConductiveElectricalNode from;
+        public ConductiveElectricalNode to;
+
+        public NodeEventArgs(ConductiveElectricalNode from, ConductiveElectricalNode to)
+        {
+            this.from = from;
+            this.to = to;
+        }
+    }
+
+    public static event System.EventHandler<NodeEventArgs> onAddNode;
+    public static event System.EventHandler<NodeEventArgs> onRemoveNode;
+
     private new void Awake()
     {
         base.Awake();
@@ -19,9 +34,9 @@ public class ConductiveElectricalNode : ElectricalNode
         //Disallow connections btw 2 conductive items (for now)
         if (other != null && !(isConductiveItem && other.isConductiveItem))
         {
-            if (AddNeighbor(other) && other is WirePilon)
+            if (AddNeighbor(other))
             {
-                (other as WirePilon).AddConductingNode(this);
+                onAddNode?.Invoke(this, new NodeEventArgs(this, other));
             }
         }
     }
@@ -36,9 +51,9 @@ public class ConductiveElectricalNode : ElectricalNode
         ConductiveElectricalNode other = collision.gameObject.GetComponentInParent<ConductiveElectricalNode>();
         if (other != null)
         {
-            if (RemoveNeighbor(other) && other is WirePilon)
+            if (RemoveNeighbor(other))
             {
-                (other as WirePilon).RemoveConductingNode(this);
+                onRemoveNode?.Invoke(this, new NodeEventArgs(this, other));
             }
         }
     }
