+++ Slider/Assets/Scripts/Map/Factory/ElectricLineCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 82053ce22e7fd8d47a3645089ed5ee8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
