+++ Slider/Assets/Scripts/Map/Factory/PowerSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 199e2a2ca0f70db4a9f86ac7625ed5b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
